requirements.txt
discord.py 
requests 
dotenv
cloudscraper
openai
parsel